+++ src/tkc/fscript.c
@@ -39,6 +39,9 @@ static ret_t func_while(fscript_t* fscript, fscript_args_t* args, value_t* resul
 static ret_t func_until(fscript_t* fscript, fscript_args_t* args, value_t* result) {
   return RET_OK;
 }
+static ret_t func_repeat_times(fscript_t* fscript, fscript_args_t* args, value_t* result) {
+  return RET_OK;
+}
 static ret_t func_noop(fscript_t* fscript, fscript_args_t* args, value_t* result) {
   return RET_OK;
 }
@@ -366,6 +369,27 @@ static ret_t fscript_exec_if(fscript_t* fscript, fscript_func_call_t* iter, valu
   return RET_OK;
 }
 
+static ret_t fscript_exec_loop_body(fscript_t* fscript, fscript_func_call_t* iter, uint32_t start,
+                                    value_t* result, bool_t* done) {
+  uint32_t i = 1;
+  for (i = start; i < iter->args.size; i++) {
+    value_reset(result);
+    fscript_eval_arg(fscript, iter, i, result);
+    if (fscript->breaked) {
+      *done = TRUE;
+      fscript->breaked = FALSE;
+      break;
+    } else if (fscript->returned) {
+      *done = TRUE;
+    } else if (fscript->continued) {
+      fscript->continued = FALSE;
+      break;
+    }
+  }
+
+  return RET_OK;
+}
+
 static ret_t fscript_exec_while_or_until(fscript_t* fscript, fscript_func_call_t* iter,
                                          value_t* result, bool_t is_while) {
   value_t condition;
@@ -375,31 +399,33 @@ static ret_t fscript_exec_while_or_until(fscript_t* fscript, fscript_func_call_t
   fscript->loop_count++;
   value_set_int(&condition, 0);
   while (!done && fscript_eval_arg(fscript, iter, 0, &condition) == RET_OK) {
-    uint32_t i = 1;
     if (is_while ? !value_bool(&condition) : value_bool(&condition)) {
       break;
     }
-
-    for (i = 1; i < iter->args.size; i++) {
-      value_reset(result);
-      fscript_eval_arg(fscript, iter, i, result);
-      if (fscript->breaked) {
-        done = TRUE;
-        fscript->breaked = FALSE;
-        break;
-      } else if (fscript->returned) {
-        done = TRUE;
-      } else if (fscript->continued) {
-        fscript->continued = FALSE;
-        break;
-      }
-    }
+    fscript_exec_loop_body(fscript, iter, 1, result, &done);
   }
   fscript->loop_count--;
 
   return RET_OK;
 }
 
+static ret_t fscript_exec_repeat_times(fscript_t* fscript, fscript_func_call_t* iter, value_t* result) {
+  value_t v;
+  uint32_t i = 0;
+  uint32_t n = 0;
+  bool_t done = FALSE;
+  return_value_if_fail(iter->args.size > 1, RET_FAIL);
+  return_value_if_fail(fscript_eval_arg(fscript, iter, 0, &v) == RET_OK, RET_BAD_PARAMS);
+
+  n = value_int(&v);
+  fscript->loop_count++;
+  for (i = 0; i < n && !done; i++) {
+    fscript_exec_loop_body(fscript, iter, 1, result, &done);
+  }
+  fscript->loop_count--;
+
+  return RET_OK;
+}
 static ret_t fscript_exec_while(fscript_t* fscript, fscript_func_call_t* iter, value_t* result) {
   return fscript_exec_while_or_until(fscript, iter, result, TRUE);
 }
@@ -416,6 +442,8 @@ static ret_t fscript_exec_core_func(fscript_t* fscript, fscript_func_call_t* ite
     return fscript_exec_while(fscript, iter, result);
   } else if (iter->func == func_until) {
     return fscript_exec_until(fscript, iter, result);
+  } else if (iter->func == func_repeat_times) {
+    return fscript_exec_repeat_times(fscript, iter, result);
   } else if (iter->func == func_function_def) {
     return RET_OK;
   }
@@ -1055,10 +1083,6 @@ static ret_t fexpr_parse_if(fscript_parser_t* parser, fscript_func_call_t* acall
   }
 }
 
-static ret_t fexpr_parse_while(fscript_parser_t* parser, fscript_func_call_t* acall) {
-  return fexpr_parse_block(parser, acall);
-}
-
 static ret_t fexpr_parse_function(fscript_parser_t* parser, value_t* result) {
   value_t v;
   fscript_args_t* args = NULL;
@@ -1105,9 +1129,11 @@ static ret_t fexpr_parse_function(fscript_parser_t* parser, value_t* result) {
     if (fscript_parser_expect_token(parser, TOKEN_LBRACKET, "expect \"{\"") == RET_OK) {
       fexpr_parse_if(parser, acall);
     }
-  } else if ((acall->func == func_while || acall->func == func_until) && acall->args.size == 1) {
+  } else if ((acall->func == func_while || acall->func == func_until ||
+              acall->func == func_repeat_times) &&
+             acall->args.size == 1) {
     if (fscript_parser_expect_token(parser, TOKEN_LBRACKET, "expect \"{\"") == RET_OK) {
-      fexpr_parse_while(parser, acall);
+      fexpr_parse_block(parser, acall);
     }
   }
 
@@ -2333,6 +2359,7 @@ static const func_entry_t s_builtin_funcs[] = {{"func", func_function_def, 4},
                                                {"if", func_if, 3},
                                                {"while", func_while, 10},
                                                {"until", func_until, 10},
+                                               {"repeat_times", func_repeat_times, 10},
                                                {"int", func_int, 1},
                                                {"i8", func_i8, 1},
                                                {"i16", func_i16, 1},