+++ tests/fscript_test.cc
@@ -1241,7 +1241,18 @@ TEST(FExr, until_statement1) {
   value_t v;
   tk_object_t* obj = object_default_create();
 
-  fscript_eval(obj, "abc=0;while(abc>=100) {abc = abc+1};abc", &v);
+  fscript_eval(obj, "abc=0;until(abc>=100) {abc = abc+1};abc", &v);
+  ASSERT_EQ(value_int(&v), 100);
+  value_reset(&v);
+
+  TK_OBJECT_UNREF(obj);
+}
+
+TEST(FExr, repeat_times_statement1) {
+  value_t v;
+  tk_object_t* obj = object_default_create();
+
+  fscript_eval(obj, "abc=0;repeat_times(100) {abc = abc+1};abc", &v);
   ASSERT_EQ(value_int(&v), 100);
   value_reset(&v);
 
@@ -1263,7 +1274,18 @@ TEST(FExr, until_statement2) {
   value_t v;
   tk_object_t* obj = object_default_create();
 
-  fscript_eval(obj, "abc=0;while(abc>=100) {abc = abc+1;if(abc > 50) {break}};abc", &v);
+  fscript_eval(obj, "abc=0;until(abc>=100) {abc = abc+1;if(abc > 50) {break}};abc", &v);
+  ASSERT_EQ(value_int(&v), 51);
+  value_reset(&v);
+
+  TK_OBJECT_UNREF(obj);
+}
+
+TEST(FExr, repeat_times_statement2) {
+  value_t v;
+  tk_object_t* obj = object_default_create();
+
+  fscript_eval(obj, "abc=0;repeat_times(100) {abc = abc+1;if(abc > 50) {break}};abc", &v);
   ASSERT_EQ(value_int(&v), 51);
   value_reset(&v);
 
@@ -1285,7 +1307,18 @@ TEST(FExr, until_statement3) {
   value_t v;
   tk_object_t* obj = object_default_create();
 
-  fscript_eval(obj, "a=0;b=0;while(a>=100) {a=a+1;if((a > 20) && (a<=50)) {continue};b=b+1};b", &v);
+  fscript_eval(obj, "a=0;b=0;until(a>=100) {a=a+1;if((a > 20) && (a<=50)) {continue};b=b+1};b", &v);
+  ASSERT_EQ(value_int(&v), 70);
+  value_reset(&v);
+
+  TK_OBJECT_UNREF(obj);
+}
+
+TEST(FExr, repeat_times_statement3) {
+  value_t v;
+  tk_object_t* obj = object_default_create();
+
+  fscript_eval(obj, "a=0;b=0;repeat_times(100) {a=a+1;if((a > 20) && (a<=50)) {continue};b=b+1};b", &v);
   ASSERT_EQ(value_int(&v), 70);
   value_reset(&v);